Bug #54323: Production job in pipeline, explicitely add new files
During the production job, the files from the build are copied to the
target repository. The files are added with ``git add -u`` which only
updates the working tree with the index, but does not add new files.

Adding new files is important, especially when the documentation
receives new files. They would then be missed in the production job and
fail the linting for documentation publishing.