fix: Trigger Cron UCR validation only once
Part of the joinscript can set the Cron UCR variables `time` and
`command`. An error is shown in the `join.log`, as `time` and `command` are set
separately:

```
E: Time or command is missing for cron job cron/ucsschool-id-connector-listener-trash-cleaner
```

Setting both in one `ucr set` statement fixes this.

Issue univention/components/ucsschool-id-connector#88